Extract Shakefiles.ElmFormat.IntegrationTests
